third_party/blink/web_tests/ppapi/plugins/plugin-destroyed-enumerate.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<embed type="application/x-blink-deprecated-test-plugin" id=plugin>
<script>
description("Verify that enumerating the properties of a detached plugin doesn't crash, but throws.");

window.jsTestIsAsync = true;

var testObject;
function runTest() {
    testObject = plugin.testObject;
    plugin.parentNode.removeChild(plugin);
    shouldThrow("Object.getOwnPropertyNames(testObject)");
    finishJSTest();
}
window.onload = runTest;
</script>
